Added a ViewerListener mechanism on the Viewer to work in the Swing thread.
Be careful, this viewer is different from the ViewerPipeListener. The
later one is made for work from another thread than the Swing thread
(and therefore from the Viewer), and uses the GS event mechanism to
communicate.

The new listener allows to work directly in the SwingThread.

Will have to write a lot of doc on this.