The view now provides acces to its rendering surface Component.
The view is still a JPanel that you can insert inside your Swing GUI,
but, as the rendering may occur in a sub-component added to this panel,
the view provides a getComponent() method that allows access to this
sub-component.

This is used when creating sub-classes of the DefaultView that use
other rendering mechanisms, like when using OpenGL where you have to
create your own GLCanvas.