Changed the way dynamic sizes are handled in Skeletons.
The skeleton now cache the size in the units given by the user.
The size is converted on demand.

Before the skeleton cached the size in GU, but as often the size was
specified in pixels, it was necessary to recompute all the values each
time the graph bounds changed !

The rule now is : the values are converted when needed.